Removing quickwit-actors dependency to quickwit-proto
