+++ quickwit/quickwit-actors/Cargo.toml
@@ -24,7 +24,6 @@ serde_json = { workspace = true }
 sync_wrapper = { workspace = true }
 
 quickwit-common = { workspace = true }
-quickwit-proto = { workspace = true }
 
 [features]
 testsuite = []
