+++ quickwit/quickwit-proto/Cargo.toml
@@ -14,6 +14,7 @@ anyhow = { workspace = true }
 opentelemetry = { workspace = true }
 prost = { workspace = true }
 prost-types = { workspace = true }
+quickwit-actors = { workspace = true }
 quickwit-query = { workspace = true }
 tonic = { workspace = true }
 tracing = { workspace = true }
